refactor(console): reorganize sie code files
