--- packages/console/src/pages/SignInExperience/tabs/OthersTab.tsx
@@ -1,13 +1,0 @@
-import AuthenticationForm from '../components/AuthenticationForm';
-import LanguagesForm from '../components/LanguagesForm';
-import TermsForm from '../components/TermsForm';
-
-const OthersTab = () => (
-  <>
-    <TermsForm />
-    <LanguagesForm isManageLanguageVisible />
-    <AuthenticationForm />
-  </>
-);
-
-export default OthersTab;
