+++ packages/console/src/pages/SignInExperience/tabs/OthersTab/AuthenticationForm.tsx
@@ -4,8 +4,8 @@ import { useTranslation } from 'react-i18next';
 import FormField from '@/components/FormField';
 import Switch from '@/components/Switch';
 
-import type { SignInExperienceForm } from '../types';
-import * as styles from './index.module.scss';
+import type { SignInExperienceForm } from '../../types';
+import * as styles from '../index.module.scss';
 
 const AuthenticationForm = () => {
   const { t } = useTranslation(undefined, { keyPrefix: 'admin_console' });
