+++ packages/console/src/pages/SignInExperience/tabs/OthersTab/components/ManageLanguage/LanguageEditor/EditSection.module.scss
