+++ packages/console/src/pages/SignInExperience/tabs/OthersTab/components/ManageLanguage/LanguageEditor/LanguageDetails.module.scss
