+++ packages/console/src/pages/SignInExperience/tabs/OthersTab/components/ManageLanguage/LanguageEditor/LanguageItem.module.scss
