+++ packages/console/src/pages/SignInExperience/tabs/OthersTab/components/ManageLanguage/LanguageEditor/LanguageNav.module.scss
