+++ packages/console/src/pages/SignInExperience/tabs/OthersTab/components/ManageLanguage/LanguageEditor/index.module.scss
