+++ packages/console/src/pages/SignInExperience/tabs/OthersTab/index.tsx
@@ -1,0 +1,13 @@
+import AuthenticationForm from './AuthenticationForm';
+import LanguagesForm from './LanguagesForm';
+import TermsForm from './TermsForm';
+
+const OthersTab = () => (
+  <>
+    <TermsForm />
+    <LanguagesForm isManageLanguageVisible />
+    <AuthenticationForm />
+  </>
+);
+
+export default OthersTab;
