--- packages/console/src/pages/SignInExperience/tabs/SignUpAndSignInTab/components/SignInMethodEditBox/types.ts
@@ -1,3 +1,0 @@
-import type { SignInExperience } from '@logto/schemas';
-
-export type SignInMethod = SignInExperience['signIn']['methods'][number];
