--- packages/console/src/pages/SignInExperience/tabs/SignUpAndSignInTab/components/SignUpAndSignInDiffSection/types.ts
@@ -1,6 +1,0 @@
-import type { SignInIdentifier } from '@logto/schemas';
-
-export type SignInMethodsObject = Record<
-  SignInIdentifier,
-  { password: boolean; verificationCode: boolean }
->;
