+++ packages/console/src/pages/SignInExperience/types.ts
@@ -1,6 +1,13 @@
-import type { SignInExperience, SignUp } from '@logto/schemas';
+import type { SignInExperience, SignInIdentifier, SignUp } from '@logto/schemas';
 
-export type SignInExperienceForm = Omit<SignInExperience, 'signInMethods' | 'signUp'> & {
+export type SignInExperienceForm = Omit<SignInExperience, 'signUp'> & {
   signUp?: SignUp;
   createAccountEnabled: boolean;
 };
+
+export type SignInMethod = SignInExperience['signIn']['methods'][number];
+
+export type SignInMethodsObject = Record<
+  SignInIdentifier,
+  { password: boolean; verificationCode: boolean }
+>;
